backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.11.7
pydantic-settings>=2.3.4
sqlalchemy[asyncio]>=2.0.31
asyncpg>=0.29.0
greenlet>=3.0.3
httpx>=0.27.0
redis>=5.0.7
celery[redis]>=5.4.0
PyJWT>=2.8.0
python-json-logger>=2.0.7
catboost>=1.2.5
pandas>=2.2.2
numpy>=1.26.4
